Early rearrange work to separate json logic from result output logic.
