+++ commec/tests/test_json.py
@@ -1,12 +1,13 @@
 import pytest
-from commec.config.json_io import *
-from commec.tools.search_handler import SearchToolVersion
 from dataclasses import asdict
+from commec.config.json_io import *
+from commec.config.result import *
+from commec.tools.search_handler import SearchToolVersion
 
 @pytest.fixture
 def test_screendata():
-    '''Fixture to provide the ScreenData for testing.'''
-    return ScreenData(
+    '''Fixture to provide the ScreenResult for testing.'''
+    return ScreenResult(
         #recommendation="PASS",
         commec_info = CommecRunInformation(
             commec_version="0.1.2",
@@ -21,7 +22,7 @@ def test_screendata():
             date_run="1.1.2024",
         ),
         queries= [
-            QueryData(
+            QueryResult(
                 query="Query1",
                 length=10,
                 sequence="ABCDEFGHIJ",
@@ -49,12 +50,12 @@ def test_screendata():
 
 @pytest.fixture
 def empty_screendata():
-    '''Fixture to provide the ScreenData for testing.'''
-    return ScreenData()
+    '''Fixture to provide the ScreenResult for testing.'''
+    return ScreenResult()
 
 @pytest.mark.parametrize("test_data_fixture",["test_screendata", "empty_screendata"])
 def test_json_io(tmp_path, request, test_data_fixture):
-    ''' Test to ensure that read/write for JSON ScreenData I/O is working correctly.'''
+    ''' Test to ensure that read/write for JSON ScreenResult I/O is working correctly.'''
     test_data = request.getfixturevalue(test_data_fixture)
     json_filename1 = tmp_path / "testread1.json"
     json_filename2 = tmp_path / "testread2.json"
@@ -78,7 +79,7 @@ def test_json_io(tmp_path, request, test_data_fixture):
     )
 
 def test_erroneous_info(tmp_path, test_screendata):
-    ''' Test to ensure that read/write for JSON ScreenData I/O is working correctly.'''
+    ''' Test to ensure that read/write for JSON ScreenResult I/O is working correctly.'''
     test_data = test_screendata
     json_filename3 = tmp_path / "testread3.json"
     json_filename4 = tmp_path / "testread4.json"
@@ -111,11 +112,11 @@ def test_adding_data_to_existing():
     """
     Tests to ensure the mutability of writing to queries is working as expected.
     """
-    def write_info(input_query : QueryData):
+    def write_info(input_query : QueryResult):
         input_query.recommendation.biorisk_screen = CommecRecommendation.PASS
     
-    new_screen_data = ScreenData()
-    new_screen_data.queries.append(QueryData("test01", 10, "ATGCATGCAT", CommecRecommendation.FLAG))
+    new_screen_data = ScreenResult()
+    new_screen_data.queries.append(QueryResult("test01", 10, "ATGCATGCAT", CommecRecommendation.FLAG))
     write_query = new_screen_data.get_query("test01")
     write_info(write_query)
     assert new_screen_data.queries[0].recommendation.biorisk_screen == CommecRecommendation.PASS
