+++ commec/utils/json_html_output.py
@@ -1,5 +1,5 @@
 """
-Used to convert a json object created from Commec Screen, or a ScreenData object,
+Used to convert a json object created from Commec Screen, or a ScreenResult object,
 into a visual HTML representation of the Commec output. Which can be embedded into
 any other HTML document as appropriate.
 """
@@ -10,10 +10,10 @@ import plotly.graph_objects as go
 import pandas as pd
 from mako.template import Template
 
-from commec.config.json_io import (
-    get_screen_data_from_json,
-    ScreenData,
-    QueryData,
+from commec.config.json_io import get_screen_data_from_json
+from commec.config.result import (
+    ScreenResult,
+    QueryResult,
     HitDescription,
     CommecScreenStep,
 )
@@ -57,9 +57,9 @@ def color_from_hit(hit : HitDescription) -> CommecPalette:
         return CommecPalette.YELLOW
     return CommecPalette.DK_BLUE
 
-def generate_html_from_screen_data(input_data : ScreenData, output_file : str):
+def generate_html_from_screen_data(input_data : ScreenResult, output_file : str):
     """
-    Interpret the ScreenData from Commec Screen output as a visualisation, in 
+    Interpret the ScreenResult from Commec Screen output as a visualisation, in 
     the form on an HTML output.
     If Commec Screen handled multiple Queries, then they are combined in the HTML output.
     """
@@ -85,7 +85,7 @@ def generate_html_from_screen_data(input_data : ScreenData, output_file : str):
         output_file.write(rendered_html)
 
 
-def update_layout(fig, query_to_draw : QueryData, stacks):
+def update_layout(fig, query_to_draw : QueryResult, stacks):
     """ 
     Applies some default settings to the plotly figure,
     also adjusts the figure height based on the number of vertically stacking bars.
@@ -126,7 +126,7 @@ def update_layout(fig, query_to_draw : QueryData, stacks):
         'bargap': 0.01,
     })
 
-def generate_outcome_string(query : QueryData, hit : HitDescription) -> str:
+def generate_outcome_string(query : QueryResult, hit : HitDescription) -> str:
     """
     Takes a Query, and associated hit, and formats a human readable output string,
     handling associated construction logic.
@@ -190,11 +190,11 @@ def generate_outcome_string(query : QueryData, hit : HitDescription) -> str:
             #Best match to regulated viruses. 2 best match hits to nucleotides found in 1 species, 2 with regulated pathogen taxId in lineage (Influenza A)
         return "No Annotations."
 
-def draw_query_to_plot(fig : go.Figure, query_to_draw : QueryData):
+def draw_query_to_plot(fig : go.Figure, query_to_draw : QueryResult):
     """ 
     Write the data from a single query into the figure for plotly. 
     """
-    # Interpret the QueryData into bars for the plot.
+    # Interpret the QueryResult into bars for the plot.
     graph_data = [
         {"label": query_to_draw.query[:25], 
          "label_verbose": query_to_draw.query, 
@@ -304,7 +304,7 @@ def generate_html_from_screen_json(input_file : str, output_file : str):
     """ 
     Wrapper for input filepath, rather than screen data object.
     """
-    input_data : ScreenData = get_screen_data_from_json(input_file)
+    input_data : ScreenResult = get_screen_data_from_json(input_file)
     generate_html_from_screen_data(input_data, output_file)
 
 def main():
